+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.1 (unreleased)
 ----------------
 
+- Inject JS via viewlet to fix base tag to match window.location.href.
+  [seanupton]
+
 - Created policy product.
   [seanupton]